requirements.txt
Backtesting==0.3.3
matplotlib==3.8.3
numpy==1.24.1
pandas==1.5.3
pandas_ta==0.3.14b0
pillow==10.2.0
requests==2.28.2
seaborn==0.13.2
selenium==4.15.2
streamlit==1.29.0
virtualenv==20.25.1
webdriver-manager==4.0.1
yfinance==0.2.9